+++ CHANGELOG.md
@@ -8,11 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
-- Added support for AVR architecture.
 
+* Added support for AVR architecture.
 * Add Entry Api to IndexMap
 * Implement IntoIterator trait for Indexmap
 * Implement FromIterator for String
+* Add first/last API to IndexMap and IndexSet
 
 ### Changed